clean up envelope editor
