+++ src/game/editor/editor.cpp
@@ -332,6 +332,8 @@ ColorRGBA CEditor::GetButtonColor(const void *pID, int Checked)
 
 	switch(Checked)
 	{
+	case 8: // invisible
+		return ColorRGBA(0, 0, 0, 0);
 	case 7: // selected + game layers
 		if(UI()->HotItem() == pID)
 			return ColorRGBA(1, 0, 0, 0.4f);
@@ -403,13 +405,13 @@ int CEditor::DoButton_Editor(const void *pID, const char *pText, int Checked, co
 	return DoButton_Editor_Common(pID, pText, Checked, pRect, Flags, pToolTip);
 }
 
-int CEditor::DoButton_Env(const void *pID, const char *pText, int Checked, const CUIRect *pRect, const char *pToolTip, ColorRGBA BaseColor)
+int CEditor::DoButton_Env(const void *pID, const char *pText, int Checked, const CUIRect *pRect, const char *pToolTip, ColorRGBA BaseColor, int Corners)
 {
 	float Bright = Checked ? 1.0f : 0.5f;
 	float Alpha = UI()->HotItem() == pID ? 1.0f : 0.75f;
 	ColorRGBA Color = ColorRGBA(BaseColor.r * Bright, BaseColor.g * Bright, BaseColor.b * Bright, Alpha);
 
-	pRect->Draw(Color, IGraphics::CORNER_ALL, 3.0f);
+	pRect->Draw(Color, Corners, 3.0f);
 	UI()->DoLabel(pRect, pText, 10.f, TEXTALIGN_CENTER);
 	Checked %= 2;
 	return DoButton_Editor_Common(pID, pText, Checked, pRect, 0, pToolTip);
@@ -5141,25 +5143,11 @@ void CEditor::RenderModebar(CUIRect View)
 void CEditor::RenderStatusbar(CUIRect View)
 {
 	CUIRect Button;
-	View.VSplitRight(60.0f, &View, &Button);
-	static int s_EnvelopeButton = 0;
-	int MouseButton = DoButton_Editor(&s_EnvelopeButton, "Envelopes", m_ShowEnvelopeEditor, &Button, 0, "Toggles the envelope editor.");
-	if(MouseButton == 2)
-		m_ShowEnvelopeEditor = (m_ShowEnvelopeEditor + 3) % 4;
-	else if(MouseButton == 1)
-		m_ShowEnvelopeEditor = (m_ShowEnvelopeEditor + 1) % 4;
-
-	if(MouseButton)
-	{
-		m_ShowServerSettingsEditor = false;
-	}
-
-	View.VSplitRight(100.0f, &View, &Button);
-	Button.VSplitRight(10.0f, &Button, nullptr);
+	View.VSplitRight(90.0f, &View, &Button);
 	static int s_SettingsButton = 0;
 	if(DoButton_Editor(&s_SettingsButton, "Server settings", m_ShowServerSettingsEditor, &Button, 0, "Toggles the server settings editor."))
 	{
-		m_ShowEnvelopeEditor = 0;
+		m_ShowEnvelopeEditor = false;
 		m_ShowServerSettingsEditor ^= 1;
 	}
 
@@ -5228,544 +5216,6 @@ void CEditor::RemoveUnusedEnvelopes()
 		else
 		{
 			m_Map.DeleteEnvelope(Envelope);
-		}
-	}
-}
-
-void CEditor::RenderEnvelopeEditor(CUIRect View)
-{
-	if(m_SelectedEnvelope < 0)
-		m_SelectedEnvelope = 0;
-	if(m_SelectedEnvelope >= (int)m_Map.m_vpEnvelopes.size())
-		m_SelectedEnvelope = m_Map.m_vpEnvelopes.size() - 1;
-
-	CEnvelope *pEnvelope = nullptr;
-	if(m_SelectedEnvelope >= 0 && m_SelectedEnvelope < (int)m_Map.m_vpEnvelopes.size())
-		pEnvelope = m_Map.m_vpEnvelopes[m_SelectedEnvelope];
-
-	CUIRect ToolBar, CurveBar, ColorBar;
-	View.HSplitTop(15.0f, &ToolBar, &View);
-	View.HSplitTop(15.0f, &CurveBar, &View);
-	ToolBar.Margin(2.0f, &ToolBar);
-	CurveBar.Margin(2.0f, &CurveBar);
-
-	bool CurrentEnvelopeSwitched = false;
-
-	// do the toolbar
-	{
-		CUIRect Button;
-		CEnvelope *pNewEnv = nullptr;
-
-		ToolBar.VSplitRight(50.0f, &ToolBar, &Button);
-		static int s_NewSoundButton = 0;
-		if(DoButton_Editor(&s_NewSoundButton, "Sound+", 0, &Button, 0, "Creates a new sound envelope"))
-		{
-			m_Map.m_Modified = true;
-			pNewEnv = m_Map.NewEnvelope(1);
-		}
-
-		ToolBar.VSplitRight(5.0f, &ToolBar, nullptr);
-		ToolBar.VSplitRight(50.0f, &ToolBar, &Button);
-		static int s_New4dButton = 0;
-		if(DoButton_Editor(&s_New4dButton, "Color+", 0, &Button, 0, "Creates a new color envelope"))
-		{
-			m_Map.m_Modified = true;
-			pNewEnv = m_Map.NewEnvelope(4);
-		}
-
-		ToolBar.VSplitRight(5.0f, &ToolBar, nullptr);
-		ToolBar.VSplitRight(50.0f, &ToolBar, &Button);
-		static int s_New2dButton = 0;
-		if(DoButton_Editor(&s_New2dButton, "Pos.+", 0, &Button, 0, "Creates a new position envelope"))
-		{
-			m_Map.m_Modified = true;
-			pNewEnv = m_Map.NewEnvelope(3);
-		}
-
-		if(m_SelectedEnvelope >= 0)
-		{
-			// Delete button
-			ToolBar.VSplitRight(10.0f, &ToolBar, nullptr);
-			ToolBar.VSplitRight(25.0f, &ToolBar, &Button);
-			static int s_DeleteButton = 0;
-			if(DoButton_Editor(&s_DeleteButton, "✗", 0, &Button, 0, "Delete this envelope"))
-			{
-				m_Map.DeleteEnvelope(m_SelectedEnvelope);
-				if(m_SelectedEnvelope >= (int)m_Map.m_vpEnvelopes.size())
-					m_SelectedEnvelope = m_Map.m_vpEnvelopes.size() - 1;
-				pEnvelope = m_SelectedEnvelope >= 0 ? m_Map.m_vpEnvelopes[m_SelectedEnvelope] : nullptr;
-			}
-
-			// Move right button
-			ToolBar.VSplitRight(5.0f, &ToolBar, nullptr);
-			ToolBar.VSplitRight(25.0f, &ToolBar, &Button);
-			static int s_MoveRightButton = 0;
-			if(DoButton_Ex(&s_MoveRightButton, "→", 0, &Button, 0, "Move this envelope to the right", IGraphics::CORNER_R))
-			{
-				m_Map.SwapEnvelopes(m_SelectedEnvelope, m_SelectedEnvelope + 1);
-				m_SelectedEnvelope = clamp<int>(m_SelectedEnvelope + 1, 0, m_Map.m_vpEnvelopes.size() - 1);
-				pEnvelope = m_Map.m_vpEnvelopes[m_SelectedEnvelope];
-			}
-
-			// Move left button
-			ToolBar.VSplitRight(25.0f, &ToolBar, &Button);
-			static int s_MoveLeftButton = 0;
-			if(DoButton_Ex(&s_MoveLeftButton, "←", 0, &Button, 0, "Move this envelope to the left", IGraphics::CORNER_L))
-			{
-				m_Map.SwapEnvelopes(m_SelectedEnvelope - 1, m_SelectedEnvelope);
-				m_SelectedEnvelope = clamp<int>(m_SelectedEnvelope - 1, 0, m_Map.m_vpEnvelopes.size() - 1);
-				pEnvelope = m_Map.m_vpEnvelopes[m_SelectedEnvelope];
-			}
-
-			// Margin on the right side
-			ToolBar.VSplitRight(7.0f, &ToolBar, nullptr);
-		}
-
-		if(pNewEnv) // add the default points
-		{
-			if(pNewEnv->m_Channels == 4)
-			{
-				pNewEnv->AddPoint(0, f2fx(1.0f), f2fx(1.0f), f2fx(1.0f), f2fx(1.0f));
-				pNewEnv->AddPoint(1000, f2fx(1.0f), f2fx(1.0f), f2fx(1.0f), f2fx(1.0f));
-			}
-			else
-			{
-				pNewEnv->AddPoint(0, 0);
-				pNewEnv->AddPoint(1000, 0);
-			}
-		}
-
-		CUIRect Shifter, Inc, Dec;
-		ToolBar.VSplitLeft(60.0f, &Shifter, &ToolBar);
-		Shifter.VSplitRight(15.0f, &Shifter, &Inc);
-		Shifter.VSplitLeft(15.0f, &Dec, &Shifter);
-		char aBuf[64];
-		str_format(aBuf, sizeof(aBuf), "%d/%d", m_SelectedEnvelope + 1, (int)m_Map.m_vpEnvelopes.size());
-
-		ColorRGBA EnvColor = ColorRGBA(1, 1, 1, 0.5f);
-		if(!m_Map.m_vpEnvelopes.empty())
-		{
-			EnvColor = IsEnvelopeUsed(m_SelectedEnvelope) ?
-					   ColorRGBA(1, 0.7f, 0.7f, 0.5f) :
-					   ColorRGBA(0.7f, 1, 0.7f, 0.5f);
-		}
-
-		static int s_EnvelopeSelector = 0;
-		int NewValue = UiDoValueSelector(&s_EnvelopeSelector, &Shifter, aBuf, m_SelectedEnvelope + 1, 1, m_Map.m_vpEnvelopes.size(), 1, 1.0f, "Select Envelope", false, false, IGraphics::CORNER_NONE, &EnvColor, false);
-		if(NewValue - 1 != m_SelectedEnvelope)
-		{
-			m_SelectedEnvelope = NewValue - 1;
-			CurrentEnvelopeSwitched = true;
-		}
-
-		static int s_PrevButton = 0;
-		if(DoButton_ButtonDec(&s_PrevButton, nullptr, 0, &Dec, 0, "Previous Envelope"))
-		{
-			m_SelectedEnvelope--;
-			if(m_SelectedEnvelope < 0)
-				m_SelectedEnvelope = m_Map.m_vpEnvelopes.size() - 1;
-			CurrentEnvelopeSwitched = true;
-		}
-
-		static int s_NextButton = 0;
-		if(DoButton_ButtonInc(&s_NextButton, nullptr, 0, &Inc, 0, "Next Envelope"))
-		{
-			m_SelectedEnvelope++;
-			if(m_SelectedEnvelope >= (int)m_Map.m_vpEnvelopes.size())
-				m_SelectedEnvelope = 0;
-			CurrentEnvelopeSwitched = true;
-		}
-
-		if(pEnvelope)
-		{
-			ToolBar.VSplitLeft(15.0f, nullptr, &ToolBar);
-			ToolBar.VSplitLeft(40.0f, &Button, &ToolBar);
-			UI()->DoLabel(&Button, "Name:", 10.0f, TEXTALIGN_RIGHT);
-
-			ToolBar.VSplitLeft(3.0f, nullptr, &ToolBar);
-			ToolBar.VSplitLeft(ToolBar.w > ToolBar.h * 40 ? 80.0f : 60.0f, &Button, &ToolBar);
-
-			static float s_NameBox = 0;
-			if(DoEditBox(&s_NameBox, &Button, pEnvelope->m_aName, sizeof(pEnvelope->m_aName), 10.0f, &s_NameBox, false, IGraphics::CORNER_ALL, "The name of the selected envelope"))
-			{
-				m_Map.m_Modified = true;
-			}
-		}
-	}
-
-	bool ShowColorBar = false;
-	if(pEnvelope && pEnvelope->m_Channels == 4)
-	{
-		ShowColorBar = true;
-		View.HSplitTop(20.0f, &ColorBar, &View);
-		ColorBar.Margin(2.0f, &ColorBar);
-		RenderBackground(ColorBar, m_CheckerTexture, 16.0f, 1.0f);
-	}
-
-	RenderBackground(View, m_CheckerTexture, 32.0f, 0.1f);
-
-	if(pEnvelope)
-	{
-		static std::vector<int> s_vSelection;
-		static int s_EnvelopeEditorID = 0;
-		static int s_ActiveChannels = 0xf;
-
-		ColorRGBA aColors[] = {ColorRGBA(1, 0.2f, 0.2f), ColorRGBA(0.2f, 1, 0.2f), ColorRGBA(0.2f, 0.2f, 1), ColorRGBA(1, 1, 0.2f)};
-
-		CUIRect Button;
-
-		ToolBar.VSplitLeft(15.0f, &Button, &ToolBar);
-
-		static const char *s_aapNames[4][CEnvPoint::MAX_CHANNELS] = {
-			{"V", "", "", ""},
-			{"", "", "", ""},
-			{"X", "Y", "R", ""},
-			{"R", "G", "B", "A"},
-		};
-
-		static const char *s_aapDescriptions[4][CEnvPoint::MAX_CHANNELS] = {
-			{"Volume of the envelope", "", "", ""},
-			{"", "", "", ""},
-			{"X-axis of the envelope", "Y-axis of the envelope", "Rotation of the envelope", ""},
-			{"Red value of the envelope", "Green value of the envelope", "Blue value of the envelope", "Alpha value of the envelope"},
-		};
-
-		static int s_aChannelButtons[CEnvPoint::MAX_CHANNELS] = {0};
-		int Bit = 1;
-
-		for(int i = 0; i < CEnvPoint::MAX_CHANNELS; i++, Bit <<= 1)
-		{
-			ToolBar.VSplitLeft(15.0f, &Button, &ToolBar);
-			if(i < pEnvelope->m_Channels)
-			{
-				if(DoButton_Env(&s_aChannelButtons[i], s_aapNames[pEnvelope->m_Channels - 1][i], s_ActiveChannels & Bit, &Button, s_aapDescriptions[pEnvelope->m_Channels - 1][i], aColors[i]))
-					s_ActiveChannels ^= Bit;
-			}
-		}
-
-		// sync checkbox
-		ToolBar.VSplitLeft(15.0f, nullptr, &ToolBar);
-		ToolBar.VSplitLeft(12.0f, &Button, &ToolBar);
-		static int s_SyncButton;
-		if(DoButton_Editor(&s_SyncButton, pEnvelope->m_Synchronized ? "X" : "", 0, &Button, 0, "Synchronize envelope animation to game time (restarts when you touch the start line)"))
-			pEnvelope->m_Synchronized = !pEnvelope->m_Synchronized;
-
-		ToolBar.VSplitLeft(4.0f, nullptr, &ToolBar);
-		ToolBar.VSplitLeft(40.0f, &Button, &ToolBar);
-		UI()->DoLabel(&Button, "Sync.", 10.0f, TEXTALIGN_LEFT);
-
-		float EndTime = pEnvelope->EndTime();
-		if(EndTime < 1)
-			EndTime = 1;
-
-		pEnvelope->FindTopBottom(s_ActiveChannels);
-		float Top = pEnvelope->m_Top;
-		float Bottom = pEnvelope->m_Bottom;
-
-		if(Top < 1)
-			Top = 1;
-		if(Bottom >= 0)
-			Bottom = 0;
-
-		float TimeScale = EndTime / View.w;
-		float ValueScale = (Top - Bottom) / View.h;
-
-		if(UI()->MouseInside(&View))
-			UI()->SetHotItem(&s_EnvelopeEditorID);
-
-		if(UI()->HotItem() == &s_EnvelopeEditorID)
-		{
-			// do stuff
-			if(UI()->MouseButtonClicked(1))
-			{
-				// add point
-				int Time = (int)(((UI()->MouseX() - View.x) * TimeScale) * 1000.0f);
-				ColorRGBA Channels;
-				pEnvelope->Eval(Time / 1000.0f, Channels);
-				pEnvelope->AddPoint(Time,
-					f2fx(Channels.r), f2fx(Channels.g),
-					f2fx(Channels.b), f2fx(Channels.a));
-				m_Map.m_Modified = true;
-			}
-
-			m_ShowEnvelopePreview = SHOWENV_SELECTED;
-			m_pTooltip = "Press right mouse button to create a new point";
-		}
-
-		// render lines
-		{
-			UI()->ClipEnable(&View);
-			Graphics()->TextureClear();
-			Graphics()->LinesBegin();
-			for(int c = 0; c < pEnvelope->m_Channels; c++)
-			{
-				if(s_ActiveChannels & (1 << c))
-					Graphics()->SetColor(aColors[c].r, aColors[c].g, aColors[c].b, 1);
-				else
-					Graphics()->SetColor(aColors[c].r * 0.5f, aColors[c].g * 0.5f, aColors[c].b * 0.5f, 1);
-
-				float PrevX = 0;
-				ColorRGBA Channels;
-				pEnvelope->Eval(0.000001f, Channels);
-				float PrevValue = Channels[c];
-
-				int Steps = (int)((View.w / UI()->Screen()->w) * Graphics()->ScreenWidth());
-				for(int i = 1; i <= Steps; i++)
-				{
-					float a = i / (float)Steps;
-					pEnvelope->Eval(a * EndTime, Channels);
-					float v = Channels[c];
-					v = (v - Bottom) / (Top - Bottom);
-
-					IGraphics::CLineItem LineItem(View.x + PrevX * View.w, View.y + View.h - PrevValue * View.h, View.x + a * View.w, View.y + View.h - v * View.h);
-					Graphics()->LinesDraw(&LineItem, 1);
-					PrevX = a;
-					PrevValue = v;
-				}
-			}
-			Graphics()->LinesEnd();
-			UI()->ClipDisable();
-		}
-
-		// render curve options
-		{
-			for(int i = 0; i < (int)pEnvelope->m_vPoints.size() - 1; i++)
-			{
-				float t0 = pEnvelope->m_vPoints[i].m_Time / 1000.0f / EndTime;
-				float t1 = pEnvelope->m_vPoints[i + 1].m_Time / 1000.0f / EndTime;
-
-				CUIRect v;
-				v.x = CurveBar.x + (t0 + (t1 - t0) * 0.5f) * CurveBar.w;
-				v.y = CurveBar.y;
-				v.h = CurveBar.h;
-				v.w = CurveBar.h;
-				v.x -= v.w / 2;
-				void *pID = &pEnvelope->m_vPoints[i].m_Curvetype;
-				const char *apTypeName[] = {
-					"N", "L", "S", "F", "M"};
-				const char *pTypeName = "Invalid";
-				if(0 <= pEnvelope->m_vPoints[i].m_Curvetype && pEnvelope->m_vPoints[i].m_Curvetype < (int)std::size(apTypeName))
-					pTypeName = apTypeName[pEnvelope->m_vPoints[i].m_Curvetype];
-				if(DoButton_Editor(pID, pTypeName, 0, &v, 0, "Switch curve type"))
-					pEnvelope->m_vPoints[i].m_Curvetype = (pEnvelope->m_vPoints[i].m_Curvetype + 1) % NUM_CURVETYPES;
-			}
-		}
-
-		// render colorbar
-		if(ShowColorBar)
-		{
-			Graphics()->TextureClear();
-			Graphics()->QuadsBegin();
-			for(int i = 0; i < (int)pEnvelope->m_vPoints.size() - 1; i++)
-			{
-				float r0 = fx2f(pEnvelope->m_vPoints[i].m_aValues[0]);
-				float g0 = fx2f(pEnvelope->m_vPoints[i].m_aValues[1]);
-				float b0 = fx2f(pEnvelope->m_vPoints[i].m_aValues[2]);
-				float a0 = fx2f(pEnvelope->m_vPoints[i].m_aValues[3]);
-				float r1 = fx2f(pEnvelope->m_vPoints[i + 1].m_aValues[0]);
-				float g1 = fx2f(pEnvelope->m_vPoints[i + 1].m_aValues[1]);
-				float b1 = fx2f(pEnvelope->m_vPoints[i + 1].m_aValues[2]);
-				float a1 = fx2f(pEnvelope->m_vPoints[i + 1].m_aValues[3]);
-
-				IGraphics::CColorVertex Array[4] = {IGraphics::CColorVertex(0, r0, g0, b0, a0),
-					IGraphics::CColorVertex(1, r1, g1, b1, a1),
-					IGraphics::CColorVertex(2, r1, g1, b1, a1),
-					IGraphics::CColorVertex(3, r0, g0, b0, a0)};
-				Graphics()->SetColorVertex(Array, 4);
-
-				float x0 = pEnvelope->m_vPoints[i].m_Time / 1000.0f / EndTime;
-				float x1 = pEnvelope->m_vPoints[i + 1].m_Time / 1000.0f / EndTime;
-
-				IGraphics::CQuadItem QuadItem(ColorBar.x + x0 * ColorBar.w, ColorBar.y, (x1 - x0) * ColorBar.w, ColorBar.h);
-				Graphics()->QuadsDrawTL(&QuadItem, 1);
-			}
-			Graphics()->QuadsEnd();
-		}
-
-		// render handles
-
-		// keep track of last Env
-		static void *s_pID = nullptr;
-
-		// chars for textinput
-		static char s_aStrCurTime[32] = "0.000";
-		static char s_aStrCurValue[32] = "0.000";
-
-		if(CurrentEnvelopeSwitched)
-		{
-			s_pID = nullptr;
-
-			// update displayed text
-			str_copy(s_aStrCurTime, "0.000");
-			str_copy(s_aStrCurValue, "0.000");
-		}
-
-		{
-			int CurrentValue = 0, CurrentTime = 0;
-
-			Graphics()->TextureClear();
-			Graphics()->QuadsBegin();
-			for(int c = 0; c < pEnvelope->m_Channels; c++)
-			{
-				if(!(s_ActiveChannels & (1 << c)))
-					continue;
-
-				for(size_t i = 0; i < pEnvelope->m_vPoints.size(); i++)
-				{
-					float x0 = pEnvelope->m_vPoints[i].m_Time / 1000.0f / EndTime;
-					float y0 = (fx2f(pEnvelope->m_vPoints[i].m_aValues[c]) - Bottom) / (Top - Bottom);
-					CUIRect Final;
-					Final.x = View.x + x0 * View.w;
-					Final.y = View.y + View.h - y0 * View.h;
-					Final.x -= 2.0f;
-					Final.y -= 2.0f;
-					Final.w = 4.0f;
-					Final.h = 4.0f;
-
-					void *pID = &pEnvelope->m_vPoints[i].m_aValues[c];
-
-					if(UI()->MouseInside(&Final))
-						UI()->SetHotItem(pID);
-
-					float ColorMod = 1.0f;
-
-					if(UI()->CheckActiveItem(pID))
-					{
-						if(!UI()->MouseButton(0))
-						{
-							m_SelectedQuadEnvelope = -1;
-							m_SelectedEnvelopePoint = -1;
-
-							UI()->SetActiveItem(nullptr);
-						}
-						else
-						{
-							if(Input()->ShiftIsPressed())
-							{
-								if(i != 0)
-								{
-									if(Input()->ModifierIsPressed())
-										pEnvelope->m_vPoints[i].m_Time += (int)((m_MouseDeltaX));
-									else
-										pEnvelope->m_vPoints[i].m_Time += (int)((m_MouseDeltaX * TimeScale) * 1000.0f);
-									if(pEnvelope->m_vPoints[i].m_Time < pEnvelope->m_vPoints[i - 1].m_Time)
-										pEnvelope->m_vPoints[i].m_Time = pEnvelope->m_vPoints[i - 1].m_Time + 1;
-									if(i + 1 != pEnvelope->m_vPoints.size() && pEnvelope->m_vPoints[i].m_Time > pEnvelope->m_vPoints[i + 1].m_Time)
-										pEnvelope->m_vPoints[i].m_Time = pEnvelope->m_vPoints[i + 1].m_Time - 1;
-								}
-							}
-							else
-							{
-								if(Input()->ModifierIsPressed())
-									pEnvelope->m_vPoints[i].m_aValues[c] -= f2fx(m_MouseDeltaY * 0.001f);
-								else
-									pEnvelope->m_vPoints[i].m_aValues[c] -= f2fx(m_MouseDeltaY * ValueScale);
-							}
-
-							m_SelectedQuadEnvelope = m_SelectedEnvelope;
-							m_ShowEnvelopePreview = SHOWENV_SELECTED;
-							m_SelectedEnvelopePoint = i;
-							m_Map.m_Modified = true;
-						}
-
-						ColorMod = 100.0f;
-						Graphics()->SetColor(1, 1, 1, 1);
-					}
-					else if(UI()->HotItem() == pID)
-					{
-						if(UI()->MouseButton(0))
-						{
-							s_vSelection.clear();
-							s_vSelection.push_back(i);
-							UI()->SetActiveItem(pID);
-							// track it
-							s_pID = pID;
-						}
-
-						// remove point
-						if(UI()->MouseButtonClicked(1))
-						{
-							if(s_pID == pID)
-							{
-								s_pID = nullptr;
-
-								// update displayed text
-								str_copy(s_aStrCurTime, "0.000");
-								str_copy(s_aStrCurValue, "0.000");
-							}
-
-							pEnvelope->m_vPoints.erase(pEnvelope->m_vPoints.begin() + i);
-							m_Map.m_Modified = true;
-						}
-
-						m_ShowEnvelopePreview = SHOWENV_SELECTED;
-						ColorMod = 100.0f;
-						Graphics()->SetColor(1, 0.75f, 0.75f, 1);
-						m_pTooltip = "Left mouse to drag. Hold ctrl to be more precise. Hold shift to alter time point as well. Right click to delete.";
-					}
-
-					if(pID == s_pID && (Input()->KeyIsPressed(KEY_RETURN) || Input()->KeyIsPressed(KEY_KP_ENTER)))
-					{
-						if(i != 0)
-						{
-							pEnvelope->m_vPoints[i].m_Time = str_tofloat(s_aStrCurTime) * 1000.0f;
-
-							if(pEnvelope->m_vPoints[i].m_Time < pEnvelope->m_vPoints[i - 1].m_Time)
-								pEnvelope->m_vPoints[i].m_Time = pEnvelope->m_vPoints[i - 1].m_Time + 1;
-							if(i + 1 != pEnvelope->m_vPoints.size() && pEnvelope->m_vPoints[i].m_Time > pEnvelope->m_vPoints[i + 1].m_Time)
-								pEnvelope->m_vPoints[i].m_Time = pEnvelope->m_vPoints[i + 1].m_Time - 1;
-						}
-						else
-							pEnvelope->m_vPoints[i].m_Time = 0.0f;
-
-						str_format(s_aStrCurTime, sizeof(s_aStrCurTime), "%.3f", pEnvelope->m_vPoints[i].m_Time / 1000.0f);
-
-						pEnvelope->m_vPoints[i].m_aValues[c] = f2fx(str_tofloat(s_aStrCurValue));
-					}
-
-					if(UI()->CheckActiveItem(pID))
-					{
-						CurrentTime = pEnvelope->m_vPoints[i].m_Time;
-						CurrentValue = pEnvelope->m_vPoints[i].m_aValues[c];
-
-						// update displayed text
-						str_format(s_aStrCurTime, sizeof(s_aStrCurTime), "%.3f", CurrentTime / 1000.0f);
-						str_format(s_aStrCurValue, sizeof(s_aStrCurValue), "%.3f", fx2f(CurrentValue));
-					}
-
-					if(m_SelectedQuadEnvelope == m_SelectedEnvelope && m_SelectedEnvelopePoint == (int)i)
-						Graphics()->SetColor(1.0f, 1.0f, 1.0f, 1.0f);
-					else
-						Graphics()->SetColor(aColors[c].r * ColorMod, aColors[c].g * ColorMod, aColors[c].b * ColorMod, 1.0f);
-					IGraphics::CQuadItem QuadItem(Final.x, Final.y, Final.w, Final.h);
-					Graphics()->QuadsDrawTL(&QuadItem, 1);
-				}
-			}
-			Graphics()->QuadsEnd();
-
-			CUIRect ToolBar1;
-			CUIRect ToolBar2;
-			ToolBar.VSplitMid(&ToolBar1, &ToolBar2);
-			ToolBar1.VSplitRight(3.0f, &ToolBar1, nullptr);
-			ToolBar2.VSplitRight(3.0f, &ToolBar2, nullptr);
-			if(ToolBar.w > ToolBar.h * 21)
-			{
-				CUIRect Label1;
-				CUIRect Label2;
-
-				ToolBar1.VSplitMid(&Label1, &ToolBar1);
-				ToolBar2.VSplitMid(&Label2, &ToolBar2);
-				Label1.VSplitRight(3.0f, &Label1, nullptr);
-				Label2.VSplitRight(3.0f, &Label2, nullptr);
-
-				UI()->DoLabel(&Label1, "Value:", 10.0f, TEXTALIGN_RIGHT);
-				UI()->DoLabel(&Label2, "Time (in s):", 10.0f, TEXTALIGN_RIGHT);
-			}
-
-			static float s_ValNumber = 0;
-			DoEditBox(&s_ValNumber, &ToolBar1, s_aStrCurValue, sizeof(s_aStrCurValue), 10.0f, &s_ValNumber, false, IGraphics::CORNER_ALL, "The value of the selected envelope point");
-			static float s_TimeNumber = 0;
-			DoEditBox(&s_TimeNumber, &ToolBar2, s_aStrCurTime, sizeof(s_aStrCurTime), 10.0f, &s_TimeNumber, false, IGraphics::CORNER_ALL, "The time of the selected envelope point");
 		}
 	}
 }
@@ -6122,14 +5572,7 @@ void CEditor::Render()
 		View.HSplitBottom(16.0f, &View, &StatusBar);
 
 		if(m_ShowEnvelopeEditor && !m_ShowPicker)
-		{
-			float Size = 125.0f;
-			if(m_ShowEnvelopeEditor == 2)
-				Size *= 2.0f;
-			else if(m_ShowEnvelopeEditor == 3)
-				Size *= 3.0f;
-			View.HSplitBottom(Size, &View, &ExtraEditor);
-		}
+			View.HSplitBottom(m_EnvelopeEditorSplit, &View, &ExtraEditor);
 
 		if(m_ShowServerSettingsEditor && !m_ShowPicker)
 			View.HSplitBottom(250.0f, &View, &ExtraEditor);
