+++ src/game/editor/editor.h
@@ -840,10 +840,11 @@ public:
 		m_AnimateSpeed = 1;
 
 		m_ShowEnvelopeEditor = 0;
-		m_ShowServerSettingsEditor = false;
+		m_EnvelopeEditorSplit = 250.0f;
 		m_EnvelopeSelectorSplit = 200.0f;
 		m_LastEnvelopeSelectorSplit = 38.0f;
 		m_NewEnvelopeType = -1;
+		m_ShowServerSettingsEditor = false;
 
 		m_ShowEnvelopePreview = SHOWENV_NONE;
 		m_SelectedQuadEnvelope = -1;
@@ -1089,7 +1090,8 @@ public:
 	float m_AnimateTime;
 	float m_AnimateSpeed;
 
-	int m_ShowEnvelopeEditor;
+	bool m_ShowEnvelopeEditor;
+	float m_EnvelopeEditorSplit;
 	float m_EnvelopeSelectorSplit;
 	float m_LastEnvelopeSelectorSplit;
 	int m_NewEnvelopeType;
@@ -1146,7 +1148,7 @@ public:
 	void UpdateTooltip(const void *pID, const CUIRect *pRect, const char *pToolTip);
 	int DoButton_Editor_Common(const void *pID, const char *pText, int Checked, const CUIRect *pRect, int Flags, const char *pToolTip);
 	int DoButton_Editor(const void *pID, const char *pText, int Checked, const CUIRect *pRect, int Flags, const char *pToolTip, int AlignVert = 1);
-	int DoButton_Env(const void *pID, const char *pText, int Checked, const CUIRect *pRect, const char *pToolTip, ColorRGBA Color);
+	int DoButton_Env(const void *pID, const char *pText, int Checked, const CUIRect *pRect, const char *pToolTip, ColorRGBA Color, int Corners);
 
 	int DoButton_Tab(const void *pID, const char *pText, int Checked, const CUIRect *pRect, int Flags, const char *pToolTip);
 	int DoButton_Ex(const void *pID, const char *pText, int Checked, const CUIRect *pRect, int Flags, const char *pToolTip, int Corners, float FontSize = 10.0f, int AlignVert = 1);
